Split pre-genesis deposit handling out of StateProcessor. Don't generate multiple genesis events.
